adjusted several namespaces
