+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationBlocklistGuard/BlockedScreenParameters.cs
@@ -1,4 +1,4 @@
-﻿namespace DCL.ApplicationBlocklistGuard
+﻿namespace DCL.ApplicationsGuards
 {
     public struct BlockedScreenParameters
     {
