+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationBlocklistGuard/ModerationDataProvider.cs
@@ -3,7 +3,7 @@ using DCL.Multiplayer.Connections.DecentralandUrls;
 using DCL.WebRequests;
 using System.Threading;
 
-namespace DCL.ApplicationBlocklistGuard
+namespace DCL.ApplicationsGuards
 {
     public class ModerationDataProvider
     {
