--- Explorer/Assets/DCL/ApplicationsGuards/ApplicationLivekitGuard/LivekitHealtGuardController.cs
@@ -1,20 +1,0 @@
-using Cysharp.Threading.Tasks;
-using DCL.Utility;
-using MVC;
-using System.Threading;
-
-public class LivekitHealtGuardController : ControllerBase<LivekitHealthGuardView>
-{
-    public override CanvasOrdering.SortingLayer Layer { get; }
-
-    public LivekitHealtGuardController(ViewFactoryMethod viewFactory) : base(viewFactory) { }
-
-    protected override void OnViewInstantiated()
-    {
-        viewInstance.ExitButton.onClick.AddListener(ExitUtils.Exit);
-    }
-
-    protected override UniTask WaitForCloseIntentAsync(CancellationToken ct) =>
-        UniTask.Never(ct);
-
-}
