+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationLivekitGuard/LivekitHealthGuardView.cs
@@ -2,11 +2,14 @@ using MVC;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class LivekitHealthGuardView : ViewBase, IView
+namespace DCL.ApplicationsGuards
 {
+    public class LivekitHealthGuardView : ViewBase, IView
+    {
 
-    [field: SerializeField]
-    public Button ExitButton { get; private set; } = null!;
+        [field: SerializeField]
+        public Button ExitButton { get; private set; } = null!;
 
 
+    }
 }
