+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationMinimumSpecsGuard/Data/SpecProfile.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace DCL.ApplicationMinimumSpecsGuard
+namespace DCL.ApplicationsGuards
 {
     public enum SpecCategory
     {
