+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationMinimumSpecsGuard/UI/MinimumSpecsTablePresenter.cs
@@ -1,3 +1,4 @@
+using DCL.ApplicationsGuards;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
