+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationMinimumSpecsGuard/UI/Table/MinimumSpecsRowView.cs
@@ -1,3 +1,4 @@
+using DCL.ApplicationsGuards;
 using TMPro;
 using UnityEngine;
