+++ Explorer/Assets/DCL/ApplicationsGuards/ApplicationMinimumSpecsGuard/UI/Tests/MinimumSpecUIStressTester.cs
@@ -1,4 +1,5 @@
 ﻿#if UNITY_EDITOR
+using DCL.ApplicationsGuards;
 using System;
 using System.Collections.Generic;
 using System.Text;
